+++ container-engine-lib/lib/backend_impls/docker/docker_kurtosis_backend_module_functions.go
@@ -301,6 +301,10 @@ func (backend *DockerKurtosisBackend) DestroyModules(
 // ====================================================================================================
 // Gets modules matching the search filters, indexed by their container ID
 func (backend *DockerKurtosisBackend) getMatchingModules(ctx context.Context, filters *module.ModuleFilters) (map[string]*module.Module, error) {
+	if filters == nil {
+		filters = &module.ModuleFilters{}
+	}
+
 	moduleContainerSearchLabels := map[string]string{
 		label_key_consts.AppIDLabelKey.GetString():         label_value_consts.AppIDLabelValue.GetString(),
 		label_key_consts.ContainerTypeLabelKey.GetString(): label_value_consts.ModuleContainerTypeLabelValue.GetString(),